requirements.txt
# Core ML and CV libraries
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8 for detection
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0

# OCR and text recognition
easyocr>=1.7.0
pytesseract>=0.3.10
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# Data handling and augmentation
albumentations>=1.3.0
imgaug>=0.4.0
scikit-image>=0.21.0
pandas>=2.0.0

# Visualization and monitoring
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Utilities
pyyaml>=6.0
requests>=2.31.0